backend/requirements.txt
fastapi
uvicorn
requests
chromadb

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
psycopg2-binary>=2.9.9

# Audio processing
python-multipart>=0.0.6
openai-whisper>=20231117
httpx>=0.26.0
ollama>=0.1.7

# File handling
aiofiles>=23.2.1
python-magic>=0.4.27

# Additional utilities
pydantic-settings>=2.1.0

greenlet